# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Flexfence Backend API
  version: 0.1.0
  description: |
    MVP attendance API for Flexfence.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/events:
    get:
      tags: [Events]
      summary: List events
      operationId: listEvents
      security:
        - BusinessBearer: []
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEventsResponse"
        default:
          $ref: "#/components/responses/APIError"
    post:
      tags: [Events]
      summary: Create event
      operationId: createEvent
      security:
        - BusinessBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        default:
          $ref: "#/components/responses/APIError"
...
components:
  securitySchemes:
    BusinessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard business user token
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile end-user token
...
  responses:
    APIError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    CreateEventRequest:
      type: object
      required: [title, start_at, end_at]
      properties:
        title:
          type: string
        description:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
    ListEventsResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
    Event:
      type: object
      required: [id, organization_id, created_by_id, title, status, created_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        created_by_id:
          type: string
        title:
          type: string
        description:
          type: string
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        status:
          type: string
        created_at:
          type: string
          format: date-time
...